Inverse distance-weighting interpolation

`step_idw` creates a *specification* of a recipe step that will add a new 'idw' feature to a dataset based on the inverse distance-weighted mean of surrounding observations.

step_idw(
  recipe,
  target = NULL,
  lon = NULL,
  lat = NULL,
  role = "predictor",
  trained = FALSE,
  neighbors = NULL,
  dist_power = NULL,
  metrics = "idw",
  spatial = NULL,
  columns = NULL,
  skip = FALSE,
  id = rand_id("idw")
)

Arguments

recipe
A recipe.

target
Selector function to choose which variable will be used to create a new feature based on the inverse distance-weighted mean of surrounding observations.

lon, lat
Selector functions to indicate which variables should be used to calculate the distance measures to the closest observations.

role
role or model term created by this step, what analysis role should be assigned?. By default, the function assumes that resulting distance will be used as a predictor in a model.

trained
A logical that will be updated once the step has been trained.

neighbors
The number of closest neighbours to use in the distance weighting.

dist_power
Distance weighting power. The distance-based weights are raised to this power. Values > 1 cause very proximal observations to be weighted more highly.

metrics
Summary metrics to produce. One or multiple of c("idw", "mean", "sd", "raw).

spatial
Used internally to store an `sf` object of the spatial locations.

columns
Used internally to store the names of the spatial columns.

skip
A logical to skip training.

id
An identifier for the step. If omitted then this is generated automatically.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns echoing the values of `lat`, `lon`, `neighbors`, `dist_power`, and `id`.

Examples

library(gstat)
library(sf)

#> Linking to GEOS 3.7.2, GDAL 2.4.2, PROJ 5.2.0

library(tidymodels)

#> ── Attaching packages ────────────────────────────────────────────────────────────────────── tidymodels 0.1.0 ──

#> ✓ broom     0.5.5     ✓ rsample   0.0.6
#> ✓ dials     0.0.6     ✓ tibble    3.0.0
#> ✓ ggplot2   3.3.0     ✓ tune      0.1.0
#> ✓ infer     0.5.1     ✓ workflows 0.1.1
#> ✓ purrr     0.3.3     ✓ yardstick 0.0.6

#> ── Conflicts ───────────────────────────────────────────────────────────────────────── tidymodels_conflicts() ──
#> x purrr::discard()   masks scales::discard()
#> x dplyr::filter()    masks stats::filter()
#> x dplyr::lag()       masks stats::lag()
#> x ggplot2::margin()  masks dials::margin()
#> x purrr::set_names() masks magrittr::set_names()
#> x recipes::step()    masks stats::step()

data("meuse.all")

regr <- rand_forest(mode = "regression", min_n = 1) %>%
    set_engine("ranger")

rec <- meuse.all %>%
    select(x, y, zinc, elev, dist.m, soil) %>%
    recipe(zinc ~ .) %>%
    step_idw(lon = x, lat = y, target = zinc, neighbors = 3, dist_power = 2)