Feature selection step using a model's feature importance scores or coefficients

`step_importance` creates a *specification* of a recipe step that selects a subset of predictors based on the ranking of variable importance provided by a `parsnip` model specification and the `model` parameter

step_importance(
  recipe,
  ...,
  target = NULL,
  role = "predictor",
  trained = FALSE,
  model = NULL,
  num_comp = NULL,
  threshold = NULL,
  to_retain = NULL,
  scores = NULL,
  skip = FALSE,
  id = rand_id("importance1")
)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

target
name of response variable to use to evaluate the importance of the predictors against.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

model
A `model_spec` object from `parsnip`.

num_comp
numeric, the number of best scoring features to select.

threshold
numeric, percentile of best features to select. Note that this overrides num_comp.

to_retain
character, names of features to retain.

scores
tibble, tibble of feature importance scores.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

Value

a `step_importance` object.

Examples

library(parsnip)
library(recipes)
library(magrittr)

# load the example iris dataset
data(iris)

# define a base model to use for feature importances
base_model <- rand_forest(mode = "classification") %>%
    set_engine("ranger", importance = "permutation")

# create a preprocessing recipe
rec <- iris %>%
 recipe(Species ~ .) %>%
 step_importance(all_predictors(), model = base_model, num_comp = 2,
                 target = Species, id = "importance_filter")

prepped <- prep(rec)