Apply minimum Redundancy Maximum Relevance Feature Selection (mRMR)

`step_mrmr` creates a *specification* of a recipe step that will apply minimum Redundancy Maximum Relevance Feature Selection (mRMR) to numeric data. The top `num_comp` scoring features, or features whose scores occur in the top percentile `threshold` will be retained as new predictors.

step_mrmr(
  recipe,
  ...,
  target = NULL,
  role = NA,
  trained = FALSE,
  num_comp = NULL,
  threshold = NULL,
  threads = 0,
  to_retain = NULL,
  scores = NULL,
  skip = FALSE,
  id = rand_id("mrmr")
)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used

target
Name of response variable used to evaluate the mRMR against.

role
Not used by this step since no new variables are created

trained
A logical to indicate if the quantities for preprocessing have been estimated

num_comp
The number of best scoring features to select.

threshold
A numeric value between 0 and 1 representing the percentile of best features to select. For example, `threshold = 0.9` will retain only predictors with scores in the top 90th percentile. Note that this overrides num_comp.

threads
An integer specifying the number of threads to use for processing. The default = 0 uses all available threads.

to_retain
The names of features that will be retained. This parameter is NULL until the recipe is prepped.

scores
A tibble containing the mRMR scores of predictors. This parameter is only produced after the recipe has been trained.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

Value

a step_mrmr object.

Examples

library(recipes)
data("iris")
rec <- iris %>%
    recipe(Species ~.) %>%
    step_mrmr(all_predictors(), target = Species, num_comp = 2)
prepped <- prep(rec)
new_data <- juice(prepped)