Feature selection step using a model's feature importance scores or coefficients

`step_select_vip` creates a *specification* of a recipe step that selects a subset of predictors based on the ranking of variable importance provided by a `parsnip` model specification and the `model` parameter

step_select_vip(
  recipe,
  ...,
  outcome = NULL,
  role = "predictor",
  trained = FALSE,
  model = NULL,
  top_p = NA,
  threshold = NA,
  exclude = NULL,
  scores = NULL,
  skip = FALSE,
  id = rand_id("select_vip")
)

# S3 method for step_select_vip
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

outcome
A character string with the name of the response variable to use to calculate the feature importance scores.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

model
A `model_spec` object from `parsnip` that has a feature importances or coefficients method. The model needs to have an equivalent `pull_importances` method defined. See `?pull_importances` for how to define methods for models that are not currently supported.

top_p
An integer with the number of best scoring features to select.

threshold
A numeric value between 0 and 1 representing the percentile of best scoring features to select. Features with scores that are _larger_ than the specified threshold will be retained, for example `threshold = 0.9` will retain only predictors with scores in the top 90th percentile. Note that this overrides `top_p`.

exclude
A character vector of predictor names that will be removed from the data. This will be set when `prep()` is used on the recipe and should not be set by the user.

scores
A tibble with 'variable' and 'scores' columns containing the names of the variables and their feature importance scores. This parameter is only produced after the recipe has been trained.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

x
A `step_select_vip` object.

Value

a `step_select_vip` object.

Examples

library(parsnip)
library(recipes)

# load the example iris dataset
data(iris)

# define a base model to use for feature importances
base_model <- rand_forest(mode = "classification") %>%
    set_engine("ranger", importance = "permutation")

# create a preprocessing recipe
rec <- iris %>%
 recipe(Species ~ .) %>%
 step_select_vip(all_predictors(), model = base_model, top_p = 2,
                 outcome = "Species")

prepped <- prep(rec)

juice(prepped)

#> # A tibble: 150 x 3
#>    Petal.Length Petal.Width Species
#>           <dbl>       <dbl> <fct>  
#>  1          1.4         0.2 setosa 
#>  2          1.4         0.2 setosa 
#>  3          1.3         0.2 setosa 
#>  4          1.5         0.2 setosa 
#>  5          1.4         0.2 setosa 
#>  6          1.7         0.4 setosa 
#>  7          1.4         0.3 setosa 
#>  8          1.5         0.2 setosa 
#>  9          1.4         0.2 setosa 
#> 10          1.5         0.1 setosa 
#> # … with 140 more rows